Finished readme

+++ bill/database/migrations/1_first_migration.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE bills (
     id VARCHAR(36) PRIMARY KEY,
     currency VARCHAR(3) NOT NULL,
     status VARCHAR(20) NOT NULL,
-    total_amount DECIMAL(19,4) NOT NULL,
+    total_amount BIGINT NOT NULL,
     created_at TIMESTAMP NOT NULL,
     closed_at TIMESTAMP,
     CONSTRAINT valid_currency CHECK (currency IN ('USD', 'GEL'))
